Colaborator admin table

+++ src/app/admin/components/colaborators/colaborators.component.ts
@@ -1,9 +1,14 @@
 import { Component, OnInit } from '@angular/core';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+
 //Services
 import { CollaboratorsService } from './../../../core/service/collaborators/collaborators.service'
 
-//Models
+//>Models<
 import { Collaborator } from './../../../core/models/collaborators.model'
+
+import { MatDialog } from '@angular/material/dialog';
+import { DialogComponent } from './dialog/dialog.component'
 
 @Component({
   selector: 'app-colaborators',
@@ -17,13 +22,15 @@ export class ColaboratorsComponent implements OnInit {
 
   constructor(
     private collaboratorsService: CollaboratorsService,
+    private formBuilder: FormBuilder,
+    public dialog: MatDialog,
   ) { }
 
   ngOnInit(): void {
     this.collaboratorsService.getAllCollaborators().subscribe(resp => {
       this.colaborators = resp.map((e:any) =>{
         return {
-          id: e.payload.doc.id, 
+          id: e.payload.doc.id,
           nombre: e.payload.doc.data().nombre,
           ocupacion: e.payload.doc.data().ocupacion,
           descripcion: e.payload.doc.data().descripcion,
@@ -40,7 +47,30 @@ export class ColaboratorsComponent implements OnInit {
 
 
   delete(item:any) {
-    this.collaboratorsService.deleteCollaborator(item.id);
-    console.log("Haz eliminado a "+ item.nombre);
+    if(confirm("Estás segur@ qué deseas eliminar a "+ item.nombre)) {
+      this.collaboratorsService.deleteCollaborator(item.id);
+    }
   }
+
+  update(item:any):void {
+    let dialogRef = this.dialog.open(DialogComponent, {
+      data: {
+        nombre: item.nombre,
+        ocupacion: item.ocupacion,
+        descripcion: item.descripcion,
+        email: item.email,
+        tel: item.tel,
+        foto: item.foto,
+        idFirebase: item.idFirebase
+      }
+    })
+  }
+
+  openCreate() {
+    const dialogRef = this.dialog.open(DialogComponent);
+    console.log("abrir create")
+  }
+
+
+
 }
